backend/requirements.txt
# AutoQuant Backend Dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
